chore: update require paths after refactor
